backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
langchain>=0.1.16,<0.4
langchain-core>=0.1.45,<0.4
langchain-openai>=0.1.3,<0.4
langchain-community>=0.0.34,<0.4
langchain-openai
python-dotenv
google-api-python-client
